requirements.txt
aiofile==3.5.0
aiogram==2.11.2
aiohttp==3.7.3
aioredis==1.3.1
async-timeout==3.0.1
asyncpg==0.22.0
attrs==20.3.0
Babel==2.9.0
cachetools==4.2.1
caio==0.8.0
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
fastapi==0.63.0
google-api-core==1.26.0
google-api-python-client==1.12.8
google-auth==1.27.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.2
googleapis-common-protos==1.52.0
h11==0.12.0
hiredis==1.1.0
httpcore==0.12.3
httplib2==0.19.0
httpx==0.17.1
idna==2.10
multidict==5.1.0
oauthlib==3.1.0
packaging==20.9
protobuf==3.15.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.8.1
pyparsing==2.4.7
python-dotenv==0.15.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
rfc3986==1.4.0
rsa==4.7.1
six==1.15.0
sniffio==1.2.0
starlette==0.13.6
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.3
uvicorn==0.13.4
yarl==1.6.3
